ci: use reviewdog for CI linting
This change enables reviewdog to post ansible-lint/gitlint errors as annotations inside the Lint Action.
Ideally, this would also allow us to have the annotations in the PR review, but this is currently
not possible (see the reviewdog readme - ctrl-f "graceful degradation").
Still this is an improvement over having to sift through the defailed action log every time.