feat(pterodactyl_wings): switch to docker.com repos by default
BREAKING CHANGE

This patch changes the default docker repositories to those provided
by docker.com. This ensures that the docker version on the wings hosts
is always up-to-date and works with newer versions of the wings agent/panel.

If you'd like to continue using the distro-provided packages, set
`pterodactyl_wings_docker_source` back to `distro`.