fix: panel: only put aws backup configuration if the driver is s3
This commit removes the aws related configuration in the panel environment
variables if the selected backup driver is not `s3`.

This is useful if the user needs to specify additional `wings` related
variables, cf.
https://github.com/pterodactyl/panel/blob/1.0-develop/config/backups.php

Feels cleaner that way too: if any more variables or storage drivers emerge
in the future, this commit will ease template maintenance.